core: Add non-null annotations to avoid static analyser warnings
It is only valid to call these inline functions with non-null values.
However, static analysis may complain that the functions may dereference
the pointer incorrectly if it is only looking at the function itself
rather than including the surrounding code.

Add the appropriate annotiations to both fix warnings and improve
detection of bugs in API users.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>