Windows: Allow applications to claim all interfaces for a WINUSBX composite function
On Windows 10, when a USB composite device is using usbccgp to manage the parent device,
there can be problems when an application tries to claim multiple interfaces for a
WINSBX-enabled subfunction. For example, a common scenario is that an application
is able to claim the first interface of a given child function, but not the others.

This commit attempts to resolve this issue by making use of (potentially) available
interface association descriptors (IADs).

Within set_composite_interface(), which is normally called for the first interface
of a composite child function (Some MI_XX device, where XX is the first interface),
the set of IADs for the current configuration is retrieved and, if the current
interface matches the 'bFirstInterface' for one of the returned IADs then the API
backend, sub_api is also set up for the other 'associated' interfaces. Subsequent
calls to libusb_claim_interface for any of these 'associated' interfaces will get
properly routed to winusbx_claim_interface.

3 fields have been added to winusb_device_priv.usb_interface:
  is_associated_interface
  first_associated_interface
  num_associated_interfaces

Which are made use of by winusbx_claim_interface()--to decide whether
Initialize() or GetAssociatedInterface() function calls are needed)--
as well as within winusbx_close()--to perform proper cleanup for groups
of associated functions.

Signed-off-by: Ryan Metcalfe <ryan.metcalfe@novanta.com>