darwin: improve the error message on kernel driver detach failure
Detaching a kernel driver in macOS requires that either the binary has the
com.apple.vm.device-access entitlement or is run with effective UID 0 (root). This
commit adds an effective UID check before attempting to capture a device and prints
a warning that neither is the user root nor does the binary have the correct
entitlement.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>